+++ apps/web/src/app/[lang]/(main)/layout.tsx
@@ -15,6 +15,7 @@ import { useLocale } from "src/providers/locale";
 import { useUser } from "src/providers/user";
 import { getBaseLink } from "src/utils";
 import "./../../globals.css";
+import AuthSession from "src/providers/auth";
 
 type LayoutProps = {
   children: JSX.Element;
@@ -127,6 +128,7 @@ export default function Layout({ children }: LayoutProps) {
     },
   };
   return (
+    <AuthSession>
     <Mainlayout
       logo="https://github.com/ayasofyazilim-clomerce.png"
       title="ayasofya"
@@ -138,5 +140,6 @@ export default function Layout({ children }: LayoutProps) {
     >
       {children}
     </Mainlayout>
+    </AuthSession>
   );
 }
