+++ apps/web/src/app/[lang]/(main)/profile/page.tsx
@@ -1,8 +1,18 @@
 "use client";
+import { useSession } from "next-auth/react";
+import { useEffect } from "react";
 import { useUser } from "src/providers/user";
 
 export default function Page() {
   const { user } = useUser();
+  const session = useSession()
+  useEffect(() => {
+    console.log("Effect ");
+    async function getSession(){
+      console.log("session from effect",session)
+    }
+    getSession();
+  }, []);
 
   if (!user) return null;
